requirements.txt
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
idna==3.10
pycparser==2.22
python-nmap==0.7.1
requests==2.32.3
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
